feat: organize code in namespaces for QueryString and Paths
